edit 2-calloc.c 3

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -13,7 +13,7 @@
 void *_calloc(unsigned int nmemb, unsigned int size)
 {
 	void *p;
-	int i;
+	unsigned int i;
 	char *ptr;
 
 	if (nmemb == 0 || size == 0)
@@ -22,7 +22,7 @@ void *_calloc(unsigned int nmemb, unsigned int size)
 	ptr = (char*)p;
 	if (p == NULL)
 		return (NULL);
-	for (i = 0; i < (int)(nmemb * size); i++)
+	for (i = 0; i < (nmemb * size); i++)
 		ptr[i] = 0;
 	return (p);
 }
